fix imports and path joining
