+++ functions.py
@@ -1,11 +1,7 @@
-import subprocess
 import numpy as np
 import xarray as xr
 from luts import era_lookup, stat_vars_dict
 
-
-
-##### DATA PROCESSING FUNCTIONS #####
 
 def filter_files(files):
 # ignore Maurer files and the "diff" files
@@ -102,82 +98,3 @@ def create_empty_dataset(dict, geom_ids):
     )
 
     return ds
-
-
-
-
-##### SLURM FUNCTIONS #####
-
-def write_sbatch_head(sbatch_out_fp, conda_init_script, conda_env_name):
-    """Make a string of SBATCH commands that can be written into a .slurm script
-
-    Args:
-        conda_init_script (path_like): path to a script that contains commands for initializing the shells on the compute nodes to use conda activate
-
-    Returns:
-        sbatch_head (str): string of SBATCH commands ready to be used as parameter in sbatch-writing functions. The following gaps are left for filling with .format:
-            - output slurm filename
-    """
-    sbatch_head = (
-        "#!/bin/sh\n"
-        "#SBATCH --nodes=1\n"
-        f"#SBATCH --cpus-per-task=24\n"
-        f"#SBATCH -p t2small\n"
-        f"#SBATCH --output {sbatch_out_fp}\n"
-        # print start time
-        "echo Start slurm && date\n"
-        # prepare shell for using activate
-        f"source {conda_init_script}\n"
-        f"conda activate {conda_env_name}\n"
-    )
-
-    return sbatch_head
-
-
-def write_sbatch(
-    sbatch_fp,
-    sbatch_out_fp,
-    sbatch_head,
-    build_nc_script,
-    data_dir,
-    output_dir,
-):
-    """Write an sbatch script for building the netCDFs
-
-    Args:
-        sbatch_fp (path_like): path to .slurm script to write sbatch commands to
-        sbatch_out_fp (path_like): path to where sbatch stdout should be written
-        sbatch_head (dict): string for sbatch head script
-
-    Returns:
-        None, writes the commands to sbatch_fp
-    """
-    pycommands = "\n"
-    pycommands += (
-        f"python {build_nc_script} "
-        f"--data_dir {data_dir} "
-        f"--output_dir {output_dir} "
-    )
-    pycommands += "\n\n"
-
-    commands = sbatch_head.format(sbatch_out_fp=sbatch_out_fp) + pycommands
-
-    with open(sbatch_fp, "w") as f:
-        f.write(commands)
-
-    return
-
-
-def submit_sbatch(sbatch_fp):
-    """Submit a script to slurm via sbatch
-
-    Args:
-        sbatch_fp (pathlib.PosixPath): path to .slurm script to submit
-
-    Returns:
-        job id for submitted job
-    """
-    out = subprocess.check_output(["sbatch", str(sbatch_fp)])
-    job_id = out.decode().replace("\n", "").split(" ")[-1]
-
-    return job_id
